+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- auth: Add `OIDCAuthentifier` for OpenID Connect Authorization Code flow with
+  authlib (#76).
+
 ## [1.7.0] - 2026-05-08
 
 ### Added